// src/db/database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
// ...
      user_locations: {
        Row: {
          city: string;
          country_code: string;
          created_at: string | null;
          id: string;
          is_default: boolean | null;
          label: string | null;
          location: unknown;
          updated_at: string | null;
          user_id: string;
        };
        Insert: {
          city: string;
          country_code: string;
          created_at?: string | null;
          id?: string;
          is_default?: boolean | null;
          label?: string | null;
          location: unknown;
          updated_at?: string | null;
          user_id: string;
        };
        Update: {
          city?: string;
          country_code?: string;
          created_at?: string | null;
          id?: string;
          is_default?: boolean | null;
          label?: string | null;
          location?: unknown;
          updated_at?: string | null;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "user_locations_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
}